refactor(code_gen): move call_builtin_function into builtins/
